Mission Creation Started

+++ Assets/Scripts/Data/City.cs
@@ -8,11 +8,11 @@ public class City
     public string description;
     public List<string> interestPlaces = new List<string>();
     public List<string> museums = new List<string>();
-    public float latitude;
-    public float longitude;
+    public double latitude;
+    public double longitude;
     public string id;
 
-    public City(string name, string description, List<string> interestPlaces, List<string> museums, float latitude, float longitude, string id)
+    public City(string name, string description, List<string> interestPlaces, List<string> museums, double latitude, double longitude, string id)
     {
         this.name = name;
         this.description = description;
